Moved classes to better packages.
