+++ src/main/java/ludum/dare/actors/state/AbstractState.java
@@ -1,13 +1,15 @@
-package ludum.dare.interfaces;
+package ludum.dare.actors.state;
 
 import com.bitdecay.jump.Facing;
 import ludum.dare.components.AnimationComponent;
 import ludum.dare.components.InputComponent;
 import ludum.dare.components.PhysicsComponent;
+import ludum.dare.interfaces.IComponent;
+import ludum.dare.interfaces.IState;
 
 import java.util.Set;
 
-public abstract class AbstractState implements IState{
+public abstract class AbstractState implements IState {
     protected Set<IComponent> components;
     protected PhysicsComponent physicsComponent;
     protected AnimationComponent animationComponent;
