+++ src/main/java/ludum/dare/actors/state/JumpState.java
@@ -1,8 +1,7 @@
-package ludum.dare.actors.player;
+package ludum.dare.actors.state;
 
 import com.bitdecay.jump.control.PlayerAction;
 import ludum.dare.interfaces.IComponent;
-import ludum.dare.interfaces.AbstractState;
 import ludum.dare.interfaces.IState;
 
 import java.util.Set;
