+++ src/main/java/ludum/dare/actors/state/RunState.java
@@ -1,20 +1,19 @@
-package ludum.dare.actors.player;
+package ludum.dare.actors.state;
 
 import com.bitdecay.jump.control.PlayerAction;
 import ludum.dare.interfaces.IComponent;
-import ludum.dare.interfaces.AbstractState;
 import ludum.dare.interfaces.IState;
 
 import java.util.Set;
 
-public class StandState extends AbstractState {
+public class RunState extends AbstractState {
 
-    public StandState(Set<IComponent> components) {
+    public RunState(Set<IComponent> components) {
         super(components, null);
     }
 
     public void enter() {
-        animationComponent.animator.switchToAnimation("stand");
+        animationComponent.animator.switchToAnimation("run");
     }
 
     public void exit() {
@@ -23,7 +22,9 @@ public class StandState extends AbstractState {
 
     public IState update(float delta) {
         super.update(delta);
-        if (inputComponent.isJustPressed(PlayerAction.RIGHT) || inputComponent.isJustPressed(PlayerAction.LEFT)) return new RunState(components);
+
+        if (!inputComponent.isPressed(PlayerAction.RIGHT) &&
+                !inputComponent.isPressed(PlayerAction.LEFT)) return new StandState(components);
         else if (inputComponent.isJustPressed(PlayerAction.JUMP)) return new JumpState(components);
         return null;
     }
