+++ src/main/java/ludum/dare/actors/state/StandState.java
@@ -1,21 +1,19 @@
-package ludum.dare.actors.player;
+package ludum.dare.actors.state;
 
-import com.bitdecay.jump.Facing;
 import com.bitdecay.jump.control.PlayerAction;
 import ludum.dare.interfaces.IComponent;
-import ludum.dare.interfaces.AbstractState;
 import ludum.dare.interfaces.IState;
 
 import java.util.Set;
 
-public class RunState extends AbstractState {
+public class StandState extends AbstractState {
 
-    public RunState(Set<IComponent> components) {
+    public StandState(Set<IComponent> components) {
         super(components, null);
     }
 
     public void enter() {
-        animationComponent.animator.switchToAnimation("run");
+        animationComponent.animator.switchToAnimation("stand");
     }
 
     public void exit() {
@@ -24,9 +22,7 @@ public class RunState extends AbstractState {
 
     public IState update(float delta) {
         super.update(delta);
-
-        if (!inputComponent.isPressed(PlayerAction.RIGHT) &&
-                !inputComponent.isPressed(PlayerAction.LEFT)) return new StandState(components);
+        if (inputComponent.isJustPressed(PlayerAction.RIGHT) || inputComponent.isJustPressed(PlayerAction.LEFT)) return new RunState(components);
         else if (inputComponent.isJustPressed(PlayerAction.JUMP)) return new JumpState(components);
         return null;
     }
